+++ test/e2e_test/agent_core/runtime/test_thought_action_pause_resume.py
@@ -15,7 +15,7 @@ from gearmeshing_ai.core.database import (
     create_engine,
     create_sessionmaker,
 )
-from gearmeshing_ai.agent_core.repos.sql import build_sql_repos
+from gearmeshing_ai.core.database.repositories.bundle import build_sql_repos
 from gearmeshing_ai.agent_core.runtime import EngineDeps
 from gearmeshing_ai.agent_core.runtime.engine import AgentEngine
 from gearmeshing_ai.core.models.domain import (
@@ -38,7 +38,7 @@ async def test_mixed_thought_then_action_pause_and_resume_round_trip() -> None:
         await create_all(engine)
         session_factory = create_sessionmaker(engine)
 
-        repos = build_sql_repos(session_factory=session_factory)
+        repos = await build_sql_repos(session_factory=session_factory)
 
         async with AsyncConnectionPool(conninfo=pool_url, min_size=1, max_size=1, kwargs={"autocommit": True}) as pool:
             checkpointer = AsyncPostgresSaver(pool)
