+++ test/unit_test/server/services/test_deps.py
@@ -56,7 +56,7 @@ class TestOrchestratorDepIntegration:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -102,7 +102,7 @@ class TestOrchestratorDepIntegration:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -138,7 +138,7 @@ class TestOrchestratorDepSingleton:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -203,7 +203,7 @@ class TestOrchestratorDepWithMockFastAPI:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -243,7 +243,7 @@ class TestOrchestratorDepWithMockFastAPI:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -311,7 +311,7 @@ class TestOrchestratorDepEdgeCases:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -344,7 +344,7 @@ class TestOrchestratorDepEdgeCases:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
