+++ test/unit_test/server/services/test_orchestrator.py
@@ -39,7 +39,7 @@ class TestOrchestratorServiceInitialization:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
@@ -62,7 +62,7 @@ class TestOrchestratorServiceInitialization:
     @patch("gearmeshing_ai.server.services.orchestrator.DatabasePolicyProvider")
     @patch("gearmeshing_ai.server.services.orchestrator.StructuredPlanner")
     @patch("gearmeshing_ai.server.services.orchestrator.AgentServiceDeps")
-    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos")
+    @patch("gearmeshing_ai.server.services.orchestrator.build_sql_repos_from_session")
     @patch("gearmeshing_ai.server.services.orchestrator.AsyncPostgresSaver")
     @patch("gearmeshing_ai.server.services.orchestrator.checkpointer_pool")
     @patch("gearmeshing_ai.server.services.orchestrator.build_default_registry")
